/group/getClassStatistics、/reading-rank、/textbook-rank-by-student、阅读量都改成小数点后两位

+++ main/src/main/java/com/upc/modular/datastatistics/controller/param/StudentReadingRankParam.java
@@ -23,7 +23,7 @@ public class StudentReadingRankParam {
     private String groupName;
 
     @ApiModelProperty("阅读时长(小时)")
-    private Long readingCount;
+    private double readingCount;
 
     @ApiModelProperty("排名")
     private Long rank;
